Version 2021.2.13f1

+++ Shaders/DefaultResourcesExtra/UI/UI-DefaultETC1.shader
@@ -73,7 +73,7 @@ Shader "UI/DefaultETC1"
                 fixed4 color    : COLOR;
                 float2 texcoord  : TEXCOORD0;
                 float4 worldPosition : TEXCOORD1;
-                half4  mask : TEXCOORD2;
+                float4  mask : TEXCOORD2;
             };
 
             sampler2D _MainTex;
@@ -101,7 +101,7 @@ Shader "UI/DefaultETC1"
                 float4 clampedRect = clamp(_ClipRect, -2e10, 2e10);
                 float2 maskUV = (IN.vertex.xy - clampedRect.xy) / (clampedRect.zw - clampedRect.xy);
                 OUT.texcoord = TRANSFORM_TEX(IN.texcoord.xy, _MainTex);
-                OUT.mask = half4(IN.vertex.xy * 2 - clampedRect.xy - clampedRect.zw, 0.25 / (0.25 * half2(_UIMaskSoftnessX, _UIMaskSoftnessY) + abs(pixelSize.xy)));
+                OUT.mask = float4(IN.vertex.xy * 2 - clampedRect.xy - clampedRect.zw, 0.25 / (0.25 * half2(_UIMaskSoftnessX, _UIMaskSoftnessY) + abs(pixelSize.xy)));
 
                 OUT.color = IN.color * _Color;
                 return OUT;
